View Count Script!
I added a JavaScript script that is used to mock incrementing the view count for all my websites.

This script starts by declaring a variable count and assigns it the value of the count stored in the browser's local storage, if it exists. If the value does not exist, the variable is assigned an initial value of 884000.

It then declares a variable number and assigns it the value of the element with the id "number" in the HTML document.

It sets the text content of the number element to the current value of the count variable in a formatted string, with useGrouping: true and minimumFractionDigits: 0, this is done by using the toLocaleString() method.

It also sets the color of the text to red using the .style.color property.

The function incrementCounter() increments the count variable by 1 and updates the text content of the number element and local storage with the new count value.
The function also adds a class "updated" to the element and uses setTimeout function with a delay of 1000 milliseconds to remove the class.

The script then uses the setInterval() function to call the incrementCounter() function every 30 seconds. This means that the count variable will be incremented by 1 every 30 seconds and the number element will be updated with the new count value.

This script allows for a real-time increment of a number in a website, it also makes sure that the count value is persistent across different browser sessions by using the local storage. The toLocaleString() method is used to format the number to be more readable for the user.