Update script.js
The code above is a CSS code that is used to create animation effects for element with id "totalviewcount".

This code starts by defining a style rule for the element with the id "totalviewcount". It sets the transition property to "all" with a duration of 0.5 seconds and an easing effect of "ease". This means that any changes made to the element will have a smooth transition over 0.5 seconds.

It also sets the initial transformation of the element to "scale(1)", which means that the element will be displayed in its original size.

Then the CSS rule for the class "updated" is defined which inherits the styles of the previous rule and adds an additional property "transform: scale(1.2);" which will make the element to scale up to 120% of its original size.

This CSS code is used in conjunction with the JavaScript code to create animation effect when the element's text content is updated. The class "updated" is added to the element using JavaScript and this class triggers the animation effect.

It creates an illusion of the total view count enlarging and then returning to its original size every time the number is updated, this gives the users an indication that the view count is being updated and makes the number more noticeable.